requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
attrs==26.1.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
bidict==0.23.1
brotli==1.2.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
eval_type_backport==0.3.1
fake-http-header==0.3.5
fastapi==0.135.2
feedparser==6.0.12
ffmpy==1.0.0
filelock==3.25.2
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
ft-pandas-ta==0.3.16
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-aiplatform==1.148.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.0
gradio==6.10.0
gradio_client==2.4.0
greenlet==3.3.2
groovy==0.1.2
groq==1.1.1
grpc-google-iam-v1==0.14.4
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-gradio==0.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpointer==3.1.1
kiwisolver==1.5.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-text-splitters==1.1.1
langsmith==0.7.22
limiter==0.5.0
limits==5.8.0
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mistralai==2.1.3
mmh3==5.2.1
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
numba==0.61.2
numpy==2.2.6
ollama==0.6.1
openai==2.30.0
opentelemetry-api==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==26.0
pandas==3.0.1
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
playwright==1.58.0
postgrest==2.28.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pytz==2026.1.post1
PyYAML==6.0.3
realtime==2.28.3
regex==2026.3.32
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==65.5.0
sgmllib3k==1.0.0
shap==0.51.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slicer==0.0.8
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
ta==0.11.0
tenacity==9.1.4
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
token-bucket==0.3.0
tomlkit==0.13.3
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
vaderSentiment==3.3.2
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.1.2
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
yfinance==1.2.0
zipp==3.23.0
zstandard==0.25.0
xgboost
joblib
scikit-learn
